Further cleanup of NodeConfigEditor.
