+++ opossum_gui/src/components/node_editor/mod.rs
@@ -3,4 +3,5 @@ mod analyzer_node_editor;
 pub mod node_config_editor;
 mod optical_node_editor;
 pub use node_config_editor::NodeConfigEditor;
+pub mod hooks;
 pub mod inputs;
