+++ opossum_gui/src/components/node_editor/optical_node_editor/properties_editor/fluence_estimator_editor.rs
@@ -1,9 +1,8 @@
 use crate::components::node_editor::{
+    hooks::use_update_signal_with_reactive_prop,
     inputs::{input_components::LabeledSelect, select_options_from_enum_iterator},
     node_config_editor::NodeChangeEvent,
-    optical_node_editor::properties_editor::{
-        use_set_node_change_property, use_update_signal_with_reactive_prop,
-    },
+    optical_node_editor::properties_editor::use_set_node_change_property,
 };
 use dioxus::prelude::*;
 use inflector::Inflector;
