+++ opossum_gui/src/components/node_editor/optical_node_editor/properties_editor/light_data_editor/energy_source_editor/mod.rs
@@ -8,7 +8,7 @@ use laser_line_editor::EnergyLaserLineEditor;
 use opossum_core::prelude::{EnergyDataBuilder, LightDataBuilder};
 use spectrum_from_file_editor::SpectrumFromFileEditor;
 
-use crate::components::node_editor::optical_node_editor::properties_editor::use_update_signal_with_reactive_prop;
+use crate::components::node_editor::hooks::use_update_signal_with_reactive_prop;
 
 #[component]
 pub fn EnergySourceEditor(
