+++ opossum_gui/src/components/node_editor/optical_node_editor/properties_editor/refractive_index_editor/mod.rs
@@ -15,15 +15,14 @@ use dioxus::prelude::*;
 use inflector::Inflector;
 
 use crate::components::node_editor::{
+    hooks::use_update_signal_with_reactive_prop,
     inputs::{
         InputData, IntoInputData,
         input_components::{LabeledSelect, RowedInputs},
         select_options_from_enum_iterator,
     },
     node_config_editor::NodeChangeEvent,
-    optical_node_editor::properties_editor::{
-        use_set_node_change_property, use_update_signal_with_reactive_prop,
-    },
+    optical_node_editor::properties_editor::use_set_node_change_property,
 };
 use uuid::Uuid;
