+++ opossum_gui/src/components/node_editor/optical_node_editor/properties_editor/splitter_type_editor/mod.rs
@@ -1,11 +1,12 @@
 use crate::components::node_editor::{
     accordion::ElementList,
+    hooks::use_update_signal_with_reactive_prop,
     inputs::{input_components::LabeledSelect, select_options_from_enum_iterator},
     node_config_editor::NodeChangeEvent,
-    optical_node_editor::properties_editor::{
-        filter_type_editor::{ConstantFilterTypeEditor, SpectralFilterTypeEditor},
-        use_set_node_change_property, use_update_signal_with_reactive_prop,
+    optical_node_editor::properties_editor::filter_type_editor::{
+        ConstantFilterTypeEditor, SpectralFilterTypeEditor,
     },
+    optical_node_editor::properties_editor::use_set_node_change_property,
 };
 use dioxus::prelude::*;
 use opossum_core::{
